Review: move repository into factory package and rename to storage
