+++ cmd/agent/main.go
@@ -5,8 +5,8 @@ import (
 	"github.com/PostScripton/go-metrics-and-alerting-collection/config"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/agent"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/client"
+	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/factory/storage/memory"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/monitoring"
-	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/repository/memory"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"os"
