+++ internal/factory/storage/interfaces.go
@@ -1,4 +1,4 @@
-package repository
+package storage
 
 import (
 	"context"
